refactor: migrate project identity to com.cemc.studyabroad and update Firebase configuration credentials
