+++ android/app/build.gradle
@@ -30,7 +30,7 @@ if (keystorePropertiesFile.exists()) {
 }
 
 android {
-    namespace "com.hungdev.kltn_mobile"
+    namespace "com.cemc.studyabroad"
     compileSdk 36
     ndkVersion flutter.ndkVersion
 
@@ -55,7 +55,7 @@ android {
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
         applicationId "com.example.ably_flutter"
-        applicationId "com.hungdev.kltn_mobile"
+        applicationId "com.cemc.studyabroad"
         // You can update the following values to match your application needs.
         // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-gradle-build-configuration.
         minSdkVersion 24
