+++ android/app/src/main/kotlin/com/cemc/studyabroad/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.hungdev.kltn_mobile
+package com.cemc.studyabroad
 
 import io.flutter.embedding.android.FlutterActivity
