+++ firebase.json
@@ -1,1 +1,1 @@
-{"flutter":{"platforms":{"android":{"default":{"projectId":"kltn-56ae4","appId":"1:936640374389:android:35025d73c1f3f8c12cfa3c","fileOutput":"android/app/google-services.json"}},"dart":{"lib/firebase_options.dart":{"projectId":"kltn-56ae4","configurations":{"android":"1:936640374389:android:35025d73c1f3f8c12cfa3c","ios":"1:936640374389:ios:bbb2c5ddd56c7ddb2cfa3c","macos":"1:936640374389:ios:bbb2c5ddd56c7ddb2cfa3c","web":"1:936640374389:web:b00ba0c57f1917fa2cfa3c","windows":"1:936640374389:web:ddffa1d7cb0471d42cfa3c"}}}}}}
+{"flutter":{"platforms":{"android":{"default":{"projectId":"cemc-co","appId":"1:812956802168:android:0fc322f07d6de0cc1f618d","fileOutput":"android/app/google-services.json"}},"dart":{"lib/firebase_options.dart":{"projectId":"cemc-co","configurations":{"android":"1:812956802168:android:0fc322f07d6de0cc1f618d","ios":"1:812956802168:ios:84c2bf67303f72091f618d"}}},"ios":{"default":{"projectId":"cemc-co","appId":"1:812956802168:ios:84c2bf67303f72091f618d","uploadDebugSymbols":false,"fileOutput":"ios/Runner/GoogleService-Info.plist"}}}}}
