+++ ios/Runner/GoogleService-Info.plist
@@ -3,17 +3,17 @@
 <plist version="1.0">
 <dict>
 	<key>API_KEY</key>
-	<string>AIzaSyATFAAjvm4t57CwmPtAOS-dJjlTIek41kE</string>
+	<string>AIzaSyCwrXHri9Lns2u7my-RjqHVde3T-5TaB6s</string>
 	<key>GCM_SENDER_ID</key>
-	<string>936640374389</string>
+	<string>812956802168</string>
 	<key>PLIST_VERSION</key>
 	<string>1</string>
 	<key>BUNDLE_ID</key>
 	<string>com.hungdev.kltnMobile</string>
 	<key>PROJECT_ID</key>
-	<string>kltn-56ae4</string>
+	<string>cemc-co</string>
 	<key>STORAGE_BUCKET</key>
-	<string>kltn-56ae4.appspot.com</string>
+	<string>cemc-co.firebasestorage.app</string>
 	<key>IS_ADS_ENABLED</key>
 	<false></false>
 	<key>IS_ANALYTICS_ENABLED</key>
@@ -25,6 +25,6 @@
 	<key>IS_SIGNIN_ENABLED</key>
 	<true></true>
 	<key>GOOGLE_APP_ID</key>
-	<string>1:936640374389:ios:bbb2c5ddd56c7ddb2cfa3c</string>
+	<string>1:812956802168:ios:84c2bf67303f72091f618d</string>
 </dict>
 </plist>
