+++ ios/Runner/Info.plist
@@ -5,7 +5,7 @@
     <key>CFBundleName</key>
     <string>Canada Study Abroad</string>
     <key>CFBundleDisplayName</key>
-    <string>Study Abroad - CECM LTD.CO</string>
+    <string>Study Abroad - CEMC LTD.CO</string>
     <key>CFBundleExecutable</key>
     <string>$(EXECUTABLE_NAME)</string>
     <key>CFBundleIdentifier</key>
@@ -13,7 +13,7 @@
     <key>CFBundleInfoDictionaryVersion</key>
     <string>6.0</string>
     <key>CFBundleName</key>
-    <string>Study Abroad - CECM LTD.CO</string>
+    <string>Study Abroad - CEMC LTD.CO</string>
     <key>CFBundlePackageType</key>
     <string>APPL</string>
     <key>CFBundleShortVersionString</key>
