+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: study_abroad_cemc_mobile
-description: Study Abroad - CECM LTD.CO
+description: Multinational Study Abroad
 publish_to: "none"
 
 version: 1.10.0+10
